+++ src/main/java/school/faang/user_service/controller/event/EventParticipationController.java
@@ -1,49 +1,35 @@
 package school.faang.user_service.controller.event;
 
+import jakarta.validation.constraints.Min;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Controller;
-import school.faang.user_service.entity.User;
+import org.springframework.validation.annotation.Validated;
+import school.faang.user_service.dto.UserDto;
 import school.faang.user_service.service.event.EventParticipationService;
 
 import java.util.List;
 
 @Controller
 @RequiredArgsConstructor
+@Validated
 public class EventParticipationController {
-
-    private static final long MIN_ID_VALUE = 1;
 
     private final EventParticipationService eventParticipationService;
 
-    public void registerParticipant(long eventId, long userId) {
-        validateParticipant(eventId, userId);
+    public void registerParticipant(@Min(1) long eventId, @Min(1) long userId) {
         eventParticipationService.registerParticipant(eventId, userId);
     }
 
-    public void unregisterParticipant(long eventId, long userId) {
-        validateParticipant(eventId, userId);
+    public void unregisterParticipant(@Min(1) long eventId, @Min(1) long userId) {
         eventParticipationService.unregisterParticipant(eventId, userId);
     }
 
-    public List<User> getParticipants(long eventId) {
-        validateEvent(eventId);
+    public List<UserDto> getParticipants(@Min(1) long eventId) {
         return eventParticipationService.getParticipants(eventId);
     }
 
-    public int getParticipantsCount(long eventId) {
-        validateEvent(eventId);
+    public int getParticipantsCount(@Min(1) long eventId) {
         return eventParticipationService.getParticipantsCount(eventId);
     }
 
-    private void validateParticipant(long eventId, long userId) {
-        if (eventId < MIN_ID_VALUE || userId < MIN_ID_VALUE) {
-            throw new IllegalArgumentException("Event ID or User ID must be positive numbers.");
-        }
-    }
-
-    private void validateEvent(long eventId) {
-        if (eventId < MIN_ID_VALUE) {
-            throw new IllegalArgumentException("Event ID must be positive numbers.");
-        }
-    }
 }
