+++ src/main/java/school/faang/user_service/mapper/UserMapper.java
@@ -1,0 +1,10 @@
+package school.faang.user_service.mapper;
+
+import org.mapstruct.Mapper;
+import school.faang.user_service.dto.UserDto;
+import school.faang.user_service.entity.User;
+
+@Mapper(componentModel = "spring")
+public interface UserMapper {
+    UserDto toDto(User user);
+}
